Refactor E2E application integration test files to a singular package (#1662)
