+++ amazon-kinesis-client/src/test/java/software/amazon/kinesis/application/MultiStreamConsumerIntegrationTest.java
@@ -1,10 +1,9 @@
-package software.amazon.kinesis.lifecycle;
+package software.amazon.kinesis.application;
 
 import org.junit.Test;
-import software.amazon.kinesis.application.TestConsumer;
-import software.amazon.kinesis.config.KCLAppConfig;
-import software.amazon.kinesis.config.multistream.ReleaseCanaryMultiStreamPollingH2TestConfig;
-import software.amazon.kinesis.config.multistream.ReleaseCanaryMultiStreamStreamingTestConfig;
+import software.amazon.kinesis.application.config.KCLAppConfig;
+import software.amazon.kinesis.application.config.multistream.ReleaseCanaryMultiStreamPollingH2TestConfig;
+import software.amazon.kinesis.application.config.multistream.ReleaseCanaryMultiStreamStreamingTestConfig;
 
 public class MultiStreamConsumerIntegrationTest {
     @Test
@@ -14,7 +13,7 @@ public class MultiStreamConsumerIntegrationTest {
         consumer.run();
     }
 
-    // @Test
+    @Test
     public void kclReleaseCanaryMultiStreamStreamingTest() throws Exception {
         KCLAppConfig consumerConfig = new ReleaseCanaryMultiStreamStreamingTestConfig();
         TestConsumer consumer = new TestConsumer(consumerConfig);
