+++ amazon-kinesis-client/src/test/java/software/amazon/kinesis/application/config/ReleaseCanaryPollingH1TestConfig.java
@@ -1,4 +1,4 @@
-package software.amazon.kinesis.config;
+package software.amazon.kinesis.application.config;
 
 import java.util.Collections;
 import java.util.List;
