+++ amazon-kinesis-client/src/test/java/software/amazon/kinesis/application/config/ReleaseCanaryStreamingReshardingTestConfig.java
@@ -1,4 +1,4 @@
-package software.amazon.kinesis.config;
+package software.amazon.kinesis.application.config;
 
 import java.util.Arrays;
 import java.util.Collections;
