+++ amazon-kinesis-client/src/test/java/software/amazon/kinesis/application/config/RetrievalMode.java
@@ -1,4 +1,4 @@
-package software.amazon.kinesis.config;
+package software.amazon.kinesis.application.config;
 
 public enum RetrievalMode {
     POLLING,
