+++ amazon-kinesis-client/src/test/java/software/amazon/kinesis/application/config/crossaccount/ReleaseCanaryCrossAccountMultiStreamPollingH2TestConfig.java
@@ -1,4 +1,4 @@
-package software.amazon.kinesis.config.crossaccount;
+package software.amazon.kinesis.application.config.crossaccount;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -6,18 +6,18 @@ import java.util.List;
 import lombok.extern.slf4j.Slf4j;
 import software.amazon.awssdk.arns.Arn;
 import software.amazon.awssdk.http.Protocol;
-import software.amazon.kinesis.config.RetrievalMode;
+import software.amazon.kinesis.application.config.RetrievalMode;
 
 /**
  * Config for a cross account polling consumer with HTTP protocol of HTTP2
  */
 @Slf4j
-public class ReleaseCanaryCrossAccountMultiStreamStreamingTestConfig extends KCLCrossAccountAppConfig {
+public class ReleaseCanaryCrossAccountMultiStreamPollingH2TestConfig extends KCLCrossAccountAppConfig {
 
     private final int numStreams = 2;
-    private final String applicationName = "CrossAccountMultiStreamStreamingTest";
+    private final String applicationName = "CrossAccountMultiStreamPollingH2Test";
 
-    private final String streamName = "2XCrossAccountStreamingTestStream";
+    private final String streamName = "2XCrossAccountPollingH2TestStream";
 
     @Override
     public String getTestName() {
@@ -40,6 +40,6 @@ public class ReleaseCanaryCrossAccountMultiStreamStreamingTestConfig extends KCL
 
     @Override
     public RetrievalMode getRetrievalMode() {
-        return RetrievalMode.STREAMING;
+        return RetrievalMode.POLLING;
     }
 }
