+++ amazon-kinesis-client/src/test/java/software/amazon/kinesis/application/config/crossaccount/ReleaseCanaryCrossAccountStreamingTestConfig.java
@@ -1,4 +1,4 @@
-package software.amazon.kinesis.config.crossaccount;
+package software.amazon.kinesis.application.config.crossaccount;
 
 import java.util.Collections;
 import java.util.List;
@@ -6,17 +6,16 @@ import java.util.List;
 import lombok.extern.slf4j.Slf4j;
 import software.amazon.awssdk.arns.Arn;
 import software.amazon.awssdk.http.Protocol;
-import software.amazon.kinesis.config.RetrievalMode;
+import software.amazon.kinesis.application.config.RetrievalMode;
 
 /**
- * Config for a cross account polling consumer with HTTP protocol of HTTP2
+ * Config for a streaming consumer with HTTP protocol of HTTP2
  */
 @Slf4j
-public class ReleaseCanaryCrossAccountPollingH2TestConfig extends KCLCrossAccountAppConfig {
+public class ReleaseCanaryCrossAccountStreamingTestConfig extends KCLCrossAccountAppConfig {
 
-    private final String applicationName = "CrossAccountPollingH2Test";
-
-    private final String streamName = "CrossAccountPollingH2TestStream_";
+    private final String applicationName = "CrossAccountStreamingTest";
+    private final String streamName = "CrossAccountStreamingTestStream";
 
     @Override
     public String getTestName() {
@@ -35,6 +34,6 @@ public class ReleaseCanaryCrossAccountPollingH2TestConfig extends KCLCrossAccoun
 
     @Override
     public RetrievalMode getRetrievalMode() {
-        return RetrievalMode.POLLING;
+        return RetrievalMode.STREAMING;
     }
 }
